Refactor: give the plan/URL layer its own module
`useProductionPlan` carried ~220 lines of pure URL logic, and the token
wrapper sat in `plan-share-codec` — a module documented as a settings
codec, which owns the `s=` blob's contents but has no business owning
the whole hash. Both move to `src/lib/plan-url.ts`, which now holds the
three stages of a link end to end: token, hash body, plan state.

Folds in three things that belonged with them:

- Option defaults were written out in three places. `DEFAULT_PLAN_OPTIONS`
  is now the single declaration, and `serializeHash` omits an option by
  comparing against it rather than against a repeated literal. Those
  literals decide what an absent URL param means, so a copy drifting
  would silently change every existing link.

- `parseHash` clamps to `MAX_TARGETS`. The UI enforces it on its own add
  path, but a hash is an artifact pasted from strangers, and a crafted
  link could put the app in a state its own UI forbids.

- Preference precedence is now tested. `plan-hash.test.ts` never stubbed
  `window`, so `loadPlanOptions` always returned `{}` and every option
  assertion passed whether or not the parsed branch consulted
  preferences — the one rule whose violation silently reproduces a
  shared plan with the viewer's options. Verified the new test fails
  when that bug is simulated.

Also drops two decode paths for formats that only ever existed inside
this branch: the `1` (lz) settings-blob flag and the `tech_`-stripped
tech fallback. No commit here is on master or any tag, so no such link
can exist. That removes lz-string from the codec entirely — compression
now lives only in the hash token — and lets the blob's format check be
strict. Two tests that described those formats were rewritten to assert
what the code actually does.
