+++ src/hooks/useProductionPlan.ts
@@ -8,24 +8,13 @@ import {
   DEFAULT_MACHINES_PER_VAPORIZER,
   sanitizeMachinesPerVaporizer,
 } from "@/lib/sustain-constants";
-import {
-  loadPlanOptions,
-  savePlanOption,
-  type PlanOptions,
-} from "@/lib/plan-options-storage";
+import { savePlanOption } from "@/lib/plan-options-storage";
 import { computePlanPrune } from "@/lib/plan-prune";
 import {
-  decodeHash,
-  encodeHashToken,
   encodeSettingsSnapshot,
   withShareBlob,
 } from "@/lib/plan-share-codec";
-import {
-  decodeItemRef,
-  decodeRecipeRef,
-  encodeItemRef,
-  encodeRecipeRef,
-} from "@/lib/url-codes";
+import { encodeHashToken, parseHash, serializeHash } from "@/lib/plan-url";
 import { useTargetOptimizer } from "@/hooks/useTargetOptimizer";
 import { items, recipes, facilities, powerFuels, MAX_TARGETS } from "@/data";
 import { useState, useMemo, useCallback, useRef, useEffect } from "react";
@@ -137,228 +126,6 @@ interface SavedPlan {
  * picker + reset affordance as normal rows.
  */
 export type IneffectivePin = { itemId: ItemId; recipeId: RecipeId };
-
-/**
- * A complete plan as the URL carries it — what `parseHash` reads out and
- * what `serializeHash` writes back, so the two are provable inverses
- * (`plan-hash.test.ts` round-trips them). Naming the fields also removes
- * the swap hazard three adjacent boolean parameters used to carry.
- */
-export interface PlanHashState extends PlanOptions {
-  targets: ProductionTarget[];
-  recipeOverrides: Map<ItemId, RecipeId>;
-  manualRawMaterials: Set<ItemId>;
-}
-
-/**
- * Read a plan state out of a location hash. Defaults to the live URL;
- * the hashchange handler passes an explicit hash to decide whether an
- * externally-pasted link actually carries a plan. Exported for tests.
- *
- * Never throws — anything unparseable degrades to the target-less
- * default state.
- */
-export function parseHash(
-  rawHash: string = typeof window === "undefined" ? "" : window.location.hash,
-): PlanHashState {
-  // Stored option preferences apply ONLY on a hash-less visit ("this is
-  // my app"). A hash means "render exactly this plan state", and since
-  // `serializeHash` omits default-valued options, an absent param there
-  // means the sharer had the default — so the parsed branch below must
-  // never consult these. See `plan-options-storage.ts`.
-  const preferred = loadPlanOptions();
-  const defaultState: PlanHashState = {
-    targets: [],
-    recipeOverrides: new Map(),
-    manualRawMaterials: new Set(),
-    ceilMode: preferred.ceilMode ?? false,
-    // binFusion defaults to ON. The hash key `bf=0` opts out;
-    // omitting `bf` (or setting `bf=1`) keeps the default ON.
-    binFusion: preferred.binFusion ?? true,
-    // powerSustain defaults to OFF. The hash key `ps=1` opts in.
-    powerSustain: preferred.powerSustain ?? false,
-    // Gas-env coverage ratio defaults to 4. The hash key `mpv=N` tunes.
-    machinesPerVaporizer:
-      preferred.machinesPerVaporizer ?? DEFAULT_MACHINES_PER_VAPORIZER,
-  };
-
-  try {
-    // Unwraps the opaque token (or passes a legacy readable hash
-    // through) — see `decodeHash`.
-    const hash = decodeHash(rawHash);
-    if (!hash) return defaultState;
-
-    const params = new URLSearchParams(hash);
-
-    // Parse targets: t=item_steel:6,item_glass:3
-    // A trailing `l` on the rate marks the target as locked
-    // (t=item_steel:6l). Backward AND forward compatible: old URLs have
-    // no suffix (= unlocked), and old app versions reading a new URL
-    // still get the rate via parseFloat("6l") === 6, merely dropping
-    // the flag.
-    const targetsRaw = params.get("t");
-    const parsedTargets: ProductionTarget[] = [];
-    if (targetsRaw) {
-      for (const part of targetsRaw.split(",")) {
-        const colonIdx = part.lastIndexOf(":");
-        if (colonIdx === -1) continue;
-        const itemId = decodeItemRef(part.slice(0, colonIdx));
-        const rateStr = part.slice(colonIdx + 1);
-        const locked = rateStr.endsWith("l");
-        const rate = parseFloat(rateStr);
-        if (itemId && isFinite(rate) && rate >= 0) {
-          parsedTargets.push(
-            locked ? { itemId, rate, locked: true } : { itemId, rate },
-          );
-        }
-      }
-    }
-
-    // Parse recipeOverrides: r=<item>:<recipe> (both short codes)
-    const recipeRaw = params.get("r");
-    const parsedRecipeOverrides = new Map<ItemId, RecipeId>();
-    if (recipeRaw) {
-      for (const part of recipeRaw.split(",")) {
-        const colonIdx = part.indexOf(":");
-        if (colonIdx === -1) continue;
-        const itemId = decodeItemRef(part.slice(0, colonIdx));
-        // Resolves a code or a legacy full id, and doubles as the
-        // existence check (unknown → null → the pin is dropped).
-        const recipeId = decodeRecipeRef(part.slice(colonIdx + 1));
-        if (itemId && recipeId) {
-          parsedRecipeOverrides.set(itemId, recipeId);
-        }
-      }
-    }
-
-    // Parse manualRawMaterials: m=item_coal,item_wood
-    const manualRaw = params.get("m");
-    const parsedManualRawMaterials = new Set<ItemId>();
-    if (manualRaw) {
-      for (const rawId of manualRaw.split(",")) {
-        const itemId = decodeItemRef(rawId);
-        if (itemId) {
-          parsedManualRawMaterials.add(itemId);
-        }
-      }
-    }
-
-    // Parse ceilMode: c=1
-    const ceilRaw = params.get("c");
-    const parsedCeilMode = ceilRaw === "1";
-
-    // Parse binFusion: bf=0 disables (default on).
-    const binFusionRaw = params.get("bf");
-    const parsedBinFusion = binFusionRaw !== "0";
-
-    // Parse powerSustain: ps=1 enables (default off).
-    const parsedPowerSustain = params.get("ps") === "1";
-
-    // Parse machinesPerVaporizer: mpv=N (default 4).
-    const mpvRaw = params.get("mpv");
-    const parsedMachinesPerVaporizer =
-      mpvRaw !== null
-        ? sanitizeMachinesPerVaporizer(parseFloat(mpvRaw))
-        : DEFAULT_MACHINES_PER_VAPORIZER;
-
-    return {
-      targets: parsedTargets,
-      recipeOverrides: parsedRecipeOverrides,
-      manualRawMaterials: parsedManualRawMaterials,
-      ceilMode: parsedCeilMode,
-      binFusion: parsedBinFusion,
-      powerSustain: parsedPowerSustain,
-      machinesPerVaporizer: parsedMachinesPerVaporizer,
-    };
-  } catch {
-    return defaultState;
-  }
-}
-
-/**
- * Build the hash body for a plan state (before `encodeHashToken` wraps
- * it) — the exact inverse of `parseHash`. Exported for tests; the app
- * reaches it through the URL-sync effect and the plan-file re-entry
- * path.
- *
- * Returns `""` for a target-less plan, which keeps the URL clean.
- */
-export function serializeHash(
-  state: PlanHashState,
-  shareBlob: string,
-): string {
-  const {
-    targets,
-    recipeOverrides,
-    manualRawMaterials,
-    ceilMode,
-    binFusion,
-    powerSustain,
-    machinesPerVaporizer,
-  } = state;
-
-  // No target, no link. Options and the settings blob describe no plan
-  // on their own, so writing them would leave a long opaque hash on an
-  // empty app — and the options now survive a reload as preferences
-  // anyway (`plan-options-storage.ts`), so nothing is lost by omitting
-  // them here.
-  if (targets.length === 0) return "";
-
-  const params = new URLSearchParams();
-
-  params.set(
-    "t",
-    targets
-      .map((t) => `${encodeItemRef(t.itemId)}:${t.rate}${t.locked ? "l" : ""}`)
-      .join(","),
-  );
-
-  if (recipeOverrides.size > 0) {
-    params.set(
-      "r",
-      Array.from(recipeOverrides.entries())
-        .map(
-          ([itemId, recipeId]) =>
-            `${encodeItemRef(itemId)}:${encodeRecipeRef(recipeId)}`,
-        )
-        .join(","),
-    );
-  }
-
-  if (manualRawMaterials.size > 0) {
-    params.set("m", Array.from(manualRawMaterials, encodeItemRef).join(","));
-  }
-
-  if (ceilMode) {
-    params.set("c", "1");
-  }
-
-  // Only emit `bf=0` when the user disabled bin-fusion. The default
-  // (on) keeps the hash short.
-  if (!binFusion) {
-    params.set("bf", "0");
-  }
-
-  // Only emit `ps=1` when self-sustaining power is enabled (default off).
-  if (powerSustain) {
-    params.set("ps", "1");
-  }
-
-  // Only emit `mpv=N` when the gas-env coverage ratio differs from the
-  // default. The default keeps the hash short.
-  if (machinesPerVaporizer !== DEFAULT_MACHINES_PER_VAPORIZER) {
-    params.set("mpv", String(machinesPerVaporizer));
-  }
-
-  // `:` and `,` (the only chars URLSearchParams encodes in t/r/m — item/
-  // recipe ids are `[a-z0-9_]`, rates `[0-9.]`) are valid unencoded in a
-  // fragment, so raw-ify them for shorter, readable plan params.
-  // `parseHash` reads via `URLSearchParams.get`, which accepts both the
-  // raw and legacy `%3A`/`%2C` forms.
-  const base = params.toString().replace(/%3A/g, ":").replace(/%2C/g, ",");
-  return withShareBlob(base, shareBlob);
-}
-
 
 /**
  * Format one structured `PlanWarning` into a display string.
