+++ src/tests/lib/plan-hash.test.ts
@@ -1,27 +1,29 @@
 /**
- * Tests for `serializeHash` — the plan → hash-body writer.
+ * Tests for the plan ⇄ URL layer (`src/lib/plan-url.ts`): the token
+ * wrapper, the hash body, and the round trip between them.
  *
- * Covers the rule that a link is written ONLY for a plan that has at
- * least one target: options and the settings blob describe no plan on
- * their own, so an empty app must keep a clean, hash-less URL however
- * many toggles are flipped. The options survive a reload as preferences
- * instead (`plan-options-storage.ts`).
- *
- * The rest of the hook is not exercisable without DOM test infra, so
- * this covers the seam that is pure.
+ * Three rules carry most of the weight:
+ *   - a link is written ONLY for a plan with at least one target, so an
+ *     empty app keeps a clean URL however many toggles are flipped;
+ *   - `parseHash` and `serializeHash` are inverses, verified as a
+ *     property rather than field by field;
+ *   - stored option preferences apply on a hash-less visit and NEVER to
+ *     a link, since an absent param there means "the sharer had the
+ *     default".
  */
 
-import { describe, expect, test } from "vitest";
+import { afterEach, describe, expect, test, vi } from "vitest";
 
 import {
+  decodeHash,
+  encodeHashToken,
   parseHash,
   serializeHash,
   type PlanHashState,
-} from "@/hooks/useProductionPlan";
-import { decodeHash, encodeHashToken } from "@/lib/plan-share-codec";
-import { DEFAULT_MACHINES_PER_VAPORIZER } from "@/lib/sustain-constants";
+} from "@/lib/plan-url";
+import { DEFAULT_PLAN_OPTIONS } from "@/lib/plan-options-storage";
 import { encodeItemRef } from "@/lib/url-codes";
-import { items, recipes } from "@/data";
+import { items, recipes, MAX_TARGETS } from "@/data";
 
 const someItem = items[0].id;
 const otherItem = items[1].id;
@@ -34,10 +36,7 @@ function state(overrides: Partial<PlanHashState> = {}): PlanHashState {
     targets: [],
     recipeOverrides: new Map(),
     manualRawMaterials: new Set(),
-    ceilMode: false,
-    binFusion: true,
-    powerSustain: false,
-    machinesPerVaporizer: DEFAULT_MACHINES_PER_VAPORIZER,
+    ...DEFAULT_PLAN_OPTIONS,
     ...overrides,
   };
 }
@@ -241,5 +240,169 @@ describe("parseHash — the pasted-link guard", () => {
     expect(targetCount(`#t=${encodeItemRef(someItem)}`)).toBe(0);
     expect(targetCount(`#t=${encodeItemRef(someItem)}:abc`)).toBe(0);
     expect(targetCount(`#t=${encodeItemRef(someItem)}:-5`)).toBe(0);
+  });
+});
+
+describe("hash token (encodeHashToken / decodeHash)", () => {
+  test("round-trips an arbitrary hash body", () => {
+    const inner = "t=s:14.4l,5h:24l&c=1&s=0Dvalley~A3~a~b~c";
+    const token = encodeHashToken(inner);
+    expect(token).not.toBe(inner);
+    expect(decodeHash(token)).toBe(inner);
+    expect(decodeHash("#" + token)).toBe(inner);
+  });
+
+  test("the token stays fragment-safe and legacy-distinguishable", () => {
+    // Contract, not alphabet: either branch may win, so assert what
+    // actually matters. `=`/`&` would collide with the legacy-form
+    // detection, and anything outside the RFC 3986 fragment set would
+    // get percent-escaped on copy-paste and break the link.
+    for (const inner of [
+      "t=s:14.4l",
+      "t=s:1,5h:2,5a:3&r=s:4r&m=3a&c=1&bf=0&ps=1&mpv=6",
+      "s=0D1A3~0~1~2",
+      "t=" + "s:1,".repeat(50), // compressible → the lz branch wins
+    ]) {
+      const token = encodeHashToken(inner);
+      expect(token).toMatch(/^[01]/); // format flag
+      expect(token).toMatch(/^[A-Za-z0-9_+$-]+$/); // base64url ∪ lz-string
+      expect(token).not.toMatch(/[=&#%/]/);
+      expect(decodeHash(token)).toBe(inner);
+    }
+  });
+
+  test("picks the shorter of the two encodings", () => {
+    const compressible = "t=" + "s:1,".repeat(50);
+    expect(encodeHashToken(compressible)[0]).toBe("1");
+    // Short and high-entropy: base64url wins, since lz-string has a
+    // fixed startup cost it cannot amortize.
+    expect(encodeHashToken("t=s:14.4l")[0]).toBe("0");
+  });
+
+  test("legacy readable hashes pass through untouched (back-compat)", () => {
+    // Links shared before tokenization — every param is `k=v`, so the
+    // body always contains '='.
+    expect(decodeHash("#t=item_steel:6")).toBe("t=item_steel:6");
+    expect(decodeHash("#t=item_steel:6&c=1")).toBe("t=item_steel:6&c=1");
+    expect(decodeHash("t=item_steel:6&s=0Dvalley")).toBe(
+      "t=item_steel:6&s=0Dvalley",
+    );
+  });
+
+  test("empty in → empty out (an empty plan keeps a hash-less URL)", () => {
+    expect(encodeHashToken("")).toBe("");
+    expect(decodeHash("")).toBe("");
+    expect(decodeHash("#")).toBe("");
+  });
+
+  test("a corrupt token decodes to '' instead of throwing", () => {
+    // Truncated / mangled by a chat client, or a plain '#anchor'.
+    expect(() => decodeHash("#!!!not-base64!!!")).not.toThrow();
+    expect(decodeHash("#!!!not-base64!!!")).toBe("");
+    const token = encodeHashToken("t=s:14.4l&c=1");
+    // Any truncation must degrade, never throw.
+    for (let i = 1; i < token.length; i++) {
+      expect(() => decodeHash("#" + token.slice(0, i))).not.toThrow();
+    }
+  });
+
+  test("a token is shorter than a URL-encoded readable hash would be", () => {
+    // The point of the wrapper is cosmetic, but it must not blow the
+    // URL up either: base64url costs ~33%, which stays well under what
+    // percent-encoding the same string would cost.
+    const inner = "t=s:14.4l,5h:24l,5a:14.75&c=1&s=0Dvalley";
+    expect(encodeHashToken(inner).length).toBeLessThan(
+      encodeURIComponent(inner).length,
+    );
+  });
+});
+
+describe("option preferences apply to a hash-less visit only", () => {
+  // The rule this protects: `serializeHash` omits default-valued
+  // options, so an absent param inside a link means "the sharer had the
+  // default". Consulting the viewer's preferences on that path would
+  // reproduce a shared plan with the WRONG options — silently. Without a
+  // `window`, `loadPlanOptions` returns `{}` and this whole rule is
+  // invisible, so the fake store is the point of these tests.
+  const stubStoredOptions = (stored: Record<string, unknown>) => {
+    const store = new Map<string, string>([
+      ["endfield-calc:plan-options-v1", JSON.stringify(stored)],
+    ]);
+    vi.stubGlobal("window", {
+      localStorage: {
+        getItem: (k: string): string | null => store.get(k) ?? null,
+        setItem: (k: string, v: string): void => {
+          store.set(k, v);
+        },
+      },
+      location: { hash: "" },
+    });
+  };
+
+  afterEach(() => {
+    vi.unstubAllGlobals();
+  });
+
+  const NON_DEFAULT = {
+    ceilMode: true,
+    binFusion: false,
+    powerSustain: true,
+    machinesPerVaporizer: 9,
+  };
+
+  test("no hash → preferences win over the in-app defaults", () => {
+    stubStoredOptions(NON_DEFAULT);
+    expect(parseHash("")).toMatchObject(NON_DEFAULT);
+  });
+
+  test("a hash → the URL wins wholly, preferences are ignored", () => {
+    stubStoredOptions(NON_DEFAULT);
+    // A link whose sharer had every option at its default emits no
+    // option params at all. The result must be the defaults, NOT the
+    // viewer's stored preferences.
+    const link = encodeHashToken(
+      serializeHash(state({ targets: [{ itemId: someItem, rate: 6 }] }), BLOB),
+    );
+    expect(parseHash(`#${link}`)).toMatchObject(DEFAULT_PLAN_OPTIONS);
+  });
+
+  test("a hash overrides preferences per option, not just in bulk", () => {
+    stubStoredOptions(NON_DEFAULT);
+    const link = encodeHashToken(
+      serializeHash(
+        state({
+          targets: [{ itemId: someItem, rate: 6 }],
+          ceilMode: true, // matches the preference
+          // the other three stay default and must NOT pick up the prefs
+        }),
+        BLOB,
+      ),
+    );
+    expect(parseHash(`#${link}`)).toMatchObject({
+      ceilMode: true,
+      binFusion: DEFAULT_PLAN_OPTIONS.binFusion,
+      powerSustain: DEFAULT_PLAN_OPTIONS.powerSustain,
+      machinesPerVaporizer: DEFAULT_PLAN_OPTIONS.machinesPerVaporizer,
+    });
+  });
+
+  test("a corrupt hash falls back to preferences, not to a broken plan", () => {
+    stubStoredOptions(NON_DEFAULT);
+    expect(parseHash("#0!!!not-base64!!!")).toMatchObject(NON_DEFAULT);
+  });
+});
+
+describe("untrusted links are clamped to the UI's own ceiling", () => {
+  test("more targets than MAX_TARGETS are truncated", () => {
+    const many = Array.from({ length: MAX_TARGETS + 5 }, (_, i) => ({
+      itemId: items[i % items.length].id,
+      rate: 1,
+    }));
+    // Build the param by hand: `serializeHash` would faithfully emit all
+    // of them, and it's the READ side that has to defend the app.
+    const raw = `t=${many.map((t) => `${encodeItemRef(t.itemId)}:${t.rate}`).join(",")}`;
+    expect(parseHash(`#${encodeHashToken(raw)}`).targets.length).toBe(
+      MAX_TARGETS,
+    );
   });
 });
