+++ src/tests/lib/plan-share-codec.test.ts
@@ -26,14 +26,13 @@ import {
   type PersistedShape,
 } from "@/hooks/useDomainSettings";
 import {
-  decodeHash,
   decodeSettingsSnapshot,
-  encodeHashToken,
   encodeSettingsSnapshot,
   readShareBlobFromHash,
   shapesEqual,
   withShareBlob,
 } from "@/lib/plan-share-codec";
+import { decodeHash } from "@/lib/plan-url";
 import {
   facilities,
   items,
@@ -202,11 +201,9 @@ describe("corrupt / cross-version input", () => {
   test("empty / unrecognized-flag blobs decode to null", () => {
     expect(decodeSettingsSnapshot("")).toBeNull();
     expect(decodeSettingsSnapshot("Zbogus")).toBeNull();
-    // A legacy JSON+lz blob has no `0`/`1` flag → unrecognized → null.
-    expect(decodeSettingsSnapshot("Nbogus")).toBeNull();
-  });
-
-  test("a garbage lz (flag 1) payload never throws", () => {
+    // `0` (raw) is the only form — compression lives in the hash token —
+    // so any other leading flag is corrupt input, never a older format.
+    expect(decodeSettingsSnapshot("1@@@not-lz@@@")).toBeNull();
     expect(() => decodeSettingsSnapshot("1@@@not-lz@@@")).not.toThrow();
   });
 
@@ -241,8 +238,10 @@ describe("hash transport (withShareBlob / readShareBlobFromHash)", () => {
     expect(readShareBlobFromHash("#s=")).toBeNull();
   });
 
-  test("a '+' in the (lz) blob survives extraction", () => {
-    const blob = "1aB+c-d$eF";
+  test("the blob's `~` separators survive extraction verbatim", () => {
+    // Why the reader is a manual regex rather than `URLSearchParams`:
+    // a round trip through that would re-encode `~` and its neighbours.
+    const blob = "0D2A1~0~pump_1~1~7";
     expect(readShareBlobFromHash("#t=x&s=" + blob)).toBe(blob);
   });
 
@@ -253,80 +252,6 @@ describe("hash transport (withShareBlob / readShareBlobFromHash)", () => {
     );
     expect(extracted).toBe(encoded);
     expect(shapesEqual(decodeSettingsSnapshot(extracted!)!, MAXIMAL)).toBe(true);
-  });
-});
-
-describe("hash token (encodeHashToken / decodeHash)", () => {
-  test("round-trips an arbitrary hash body", () => {
-    const inner = "t=s:14.4l,5h:24l&c=1&s=0Dvalley~A3~a~b~c";
-    const token = encodeHashToken(inner);
-    expect(token).not.toBe(inner);
-    expect(decodeHash(token)).toBe(inner);
-    expect(decodeHash("#" + token)).toBe(inner);
-  });
-
-  test("the token stays fragment-safe and legacy-distinguishable", () => {
-    // Contract, not alphabet: either branch may win, so assert what
-    // actually matters. `=`/`&` would collide with the legacy-form
-    // detection, and anything outside the RFC 3986 fragment set would
-    // get percent-escaped on copy-paste and break the link.
-    for (const inner of [
-      "t=s:14.4l",
-      "t=s:1,5h:2,5a:3&r=s:4r&m=3a&c=1&bf=0&ps=1&mpv=6",
-      "s=0D1A3~0~1~2",
-      "t=" + "s:1,".repeat(50), // compressible → the lz branch wins
-    ]) {
-      const token = encodeHashToken(inner);
-      expect(token).toMatch(/^[01]/); // format flag
-      expect(token).toMatch(/^[A-Za-z0-9_+$-]+$/); // base64url ∪ lz-string
-      expect(token).not.toMatch(/[=&#%/]/);
-      expect(decodeHash(token)).toBe(inner);
-    }
-  });
-
-  test("picks the shorter of the two encodings", () => {
-    const compressible = "t=" + "s:1,".repeat(50);
-    expect(encodeHashToken(compressible)[0]).toBe("1");
-    // Short and high-entropy: base64url wins, since lz-string has a
-    // fixed startup cost it cannot amortize.
-    expect(encodeHashToken("t=s:14.4l")[0]).toBe("0");
-  });
-
-  test("legacy readable hashes pass through untouched (back-compat)", () => {
-    // Links shared before tokenization — every param is `k=v`, so the
-    // body always contains '='.
-    expect(decodeHash("#t=item_steel:6")).toBe("t=item_steel:6");
-    expect(decodeHash("#t=item_steel:6&c=1")).toBe("t=item_steel:6&c=1");
-    expect(decodeHash("t=item_steel:6&s=0Dvalley")).toBe(
-      "t=item_steel:6&s=0Dvalley",
-    );
-  });
-
-  test("empty in → empty out (an empty plan keeps a hash-less URL)", () => {
-    expect(encodeHashToken("")).toBe("");
-    expect(decodeHash("")).toBe("");
-    expect(decodeHash("#")).toBe("");
-  });
-
-  test("a corrupt token decodes to '' instead of throwing", () => {
-    // Truncated / mangled by a chat client, or a plain '#anchor'.
-    expect(() => decodeHash("#!!!not-base64!!!")).not.toThrow();
-    expect(decodeHash("#!!!not-base64!!!")).toBe("");
-    const token = encodeHashToken("t=s:14.4l&c=1");
-    // Any truncation must degrade, never throw.
-    for (let i = 1; i < token.length; i++) {
-      expect(() => decodeHash("#" + token.slice(0, i))).not.toThrow();
-    }
-  });
-
-  test("a token is shorter than a URL-encoded readable hash would be", () => {
-    // The point of the wrapper is cosmetic, but it must not blow the
-    // URL up either: base64url costs ~33%, which stays well under what
-    // percent-encoding the same string would cost.
-    const inner = "t=s:14.4l,5h:24l,5a:14.75&c=1&s=0Dvalley";
-    expect(encodeHashToken(inner).length).toBeLessThan(
-      encodeURIComponent(inner).length,
-    );
   });
 });
